requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
python-dotenv==1.0.1
alembic==1.13.2
gunicorn==22.0.0